sua them cac model

+++ user-service/src/main/java/com/example/recommendationservice/entity/ApDomain.java
@@ -34,7 +34,7 @@ public class ApDomain {
     private String name;
 
     @Column(name = "status")
-    private Long status;
+    private Integer status;
 
     @Column(name = "value")
     private String value;
